fix MiningIntervalPicker tests
