+++ applications/launchpad/gui-react/src/containers/MiningContainer/Statistics/MiningIntervalPicker/types.ts
@@ -1,0 +1,9 @@
+import { MiningStatisticsInterval } from '../types'
+
+export type MiningIntervalPickerComponentProps = {
+  value: Date
+  interval: MiningStatisticsInterval
+  onChange: (d: Date) => void
+  dataFrom: Date
+  dataTo: Date
+}
